+++ source/PiningSite.cpp
@@ -2,18 +2,18 @@
 
 PiningSite::PiningSite()
 {
-	position = { -1.0f, -1.0f };	//不正な値で初期化
-	pinR = -1.0f;				//不正な値で初期化
+	position = { -1.0, -1.0 };	//不正な値で初期化
+	pinR = -1.0;				//不正な値で初期化
 }
 
 PiningSite::~PiningSite(){
 	/* DO NOTHING */
 }
 
-Vector2f PiningSite::GetPinPos() {
+Vector2d PiningSite::GetPinPos() {
 	return position;
 }
 
-void PiningSite::SetPinPos(float x, float y) {
+void PiningSite::SetPinPos(double x, double y) {
 	position = { x,y };
 }
